+++ Website/src/components/Instagram/instagramContent.css
@@ -1,0 +1,61 @@
+.instagram-content {
+  padding: 2rem 7rem;
+  width: 100%;
+}
+.instagram-cards {
+  width: 100%;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  grid-template-rows: auto;
+  justify-content: center;
+  column-gap: 5rem;
+  row-gap: 2rem;
+}
+
+.instagram-card {
+  width: 100%;
+  height: 20rem;
+  background-color: #70429f;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  border-radius: 5px;
+}
+
+.instagram-card img {
+  width: 80%;
+}
+
+.instagram-stats {
+  display: flex;
+  width: 100%;
+  justify-content: space-evenly;
+}
+
+.instagram-stats > * {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.instagram-stats h5 {
+  font-size: 0.85rem;
+  font-weight: 600;
+}
+
+@media only screen and (max-width: 1261px) {
+  .instagram-cards {
+    grid-template-columns: 1fr 1fr;
+  }
+}
+@media only screen and (max-width: 835px) {
+  .instagram-cards {
+    grid-template-columns: 1fr;
+  }
+}
+@media only screen and (max-width: 472px) {
+  .instagram-content {
+    padding: 2rem 4rem;
+  }
+}
