+++ Website/src/components/Instagram/instagramData.js
@@ -1,0 +1,40 @@
+import img1 from "../../assests/images/Instagram/instagram1.jpg";
+import img2 from "../../assests/images/Instagram/instagram2.jpg";
+import img3 from "../../assests/images/Instagram/instagram3.jpg";
+import img4 from "../../assests/images/Instagram/instagram4.jpg";
+import img5 from "../../assests/images/Instagram/instagram5.jpg";
+
+const data = [
+  {
+    post: img1,
+    likes: 12,
+    comments: 23,
+    saves: 134,
+  },
+  {
+    post: img2,
+    likes: 87,
+    comments: 233,
+    saves: 9434,
+  },
+  {
+    post: img3,
+    likes: 78,
+    comments: 345,
+    saves: 245,
+  },
+  {
+    post: img4,
+    likes: 86,
+    comments: 34,
+    saves: 865,
+  },
+  {
+    post: img5,
+    likes: 92,
+    comments: 52,
+    saves: 342,
+  },
+];
+
+export default data;
