+++ Website/src/pages/Home/Home.css
@@ -1,0 +1,111 @@
+@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@200;300;400;700&display=swap");
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: "Poppins", sans-serif;
+}
+body {
+  background-color: #19161d;
+}
+
+.main {
+  height: 100%;
+  width: 100%;
+  padding: 0 7rem;
+  display: flex;
+  justify-content: space-between;
+  height: 70vh;
+}
+.main-text {
+  height: 100%;
+  font-size: 0.75rem;
+  font-weight: lighter;
+  width: 50%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  color: white;
+}
+
+.main-text h1:first-child {
+  color: #70429f;
+  font-size: 4.75rem;
+  font-weight: bold;
+}
+.main-text h1:last-child {
+  font-weight: 300;
+}
+
+.main-img {
+  height: 100%;
+  width: 50%;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.img {
+  height: 75%;
+  width: 90%;
+}
+.img img {
+  height: 95%;
+  width: 100%;
+  transform: translateX(50px);
+  opacity: 0;
+  animation: entry 6s ease infinite;
+}
+
+@keyframes entry {
+  0% {
+    transform: translateX(100px);
+    opacity: 0;
+  }
+  50% {
+    transform: translateX(0px);
+    opacity: 1;
+  }
+  100% {
+    transform: translateX(0px);
+    opacity: 1;
+  }
+}
+
+@media only screen and (max-width: 1550px) {
+  .img {
+    height: 60%;
+  }
+}
+
+@media only screen and (max-width: 1200px) {
+  .main {
+    flex-direction: column;
+    width: 100%;
+    text-align: center;
+  }
+  .main-img,
+  .main-text {
+    width: 100%;
+    height: 50vh;
+  }
+  .img img {
+    width: 60vw;
+    height: 40vh;
+    animation: none;
+    transform: translateX(0);
+    opacity: 1;
+  }
+}
+
+@media only screen and (max-width: 900px) {
+  .main {
+    padding: 0 4rem;
+    text-align: center;
+  }
+  .img img {
+    height: 25vh;
+  }
+}
